feat: add final bindef draft and format definitions
Add the BinDef reference document plus 33 format definitions. The definitions will be moved to their own
repository and provided as a compressed archive later on. The old format narrowers were removed as they are
no longer needed.

Rewrite the Result struct to be an interface instead. Each result is now its own type. To be avoid juggling
around reflections, integers are now of type big.Int rather than int64. Floats still remain as float64.

Introduce the byte[n] and signed integer format types. Evaluate the missing operatos except the comparison
operators which are currently only parsed. "Modulo" was changed to "remainder" to reflect that the sign bit
of the result is that of the dividend.