feat: add switch statements
Switch statements allow authors to define values (usually format types) based
on a series of cases. This helps avoid repetitive ifs within a structure.

Check if the identifier of a structure is empty. An empty ID means that its
value should not be included.

Clarify version addition notes in the reference docs.